Compute header statistics in a single blocked pass
update_header_stats() and validate() called min(), max(), mean() and std()
separately. That is four traversals of the data, and std() also
materialises a full-size (x - mean) temporary.

Add utils.calculate_stats(), which computes all four in one pass over
cache-sized blocks of STATS_BLOCK_SIZE elements, so peak transient memory
no longer scales with the array. Sums accumulate in float64 about a
provisional mean taken from the first block, which keeps the variance
stable for data with a large DC offset.

Behaviour change: accumulating in float64 rather than float32 can change
the stored dmean and rms by one float32 ULP, towards the more accurate
value. test_stats_are_updated_for_new_data now asserts against the float64
reference. DEVIATIONS.md records the change, the evidence for it and a
one-line revert. min and max keep the array's dtype, NaN propagates as it
does in numpy, and the validate() message text is unchanged.

+++ tests/test_mrcobject.py
@@ -418,8 +418,14 @@ class MrcObjectTest(unittest.TestCase):
         header = self.mrcobject.header
         assert header.dmin == np.float32(vol.min())
         assert header.dmax == np.float32(vol.max())
-        assert header.dmean == np.float32(vol.mean(dtype=np.float32))
-        assert header.rms == np.float32(vol.std(dtype=np.float32))
+        # FORK DEVIATION (see DEVIATIONS.md, D1). Upstream asserts bit-equality
+        # with numpy's float32-accumulated mean and std. This fork accumulates
+        # in float64, which is strictly more accurate: for this data the fork
+        # produces the nearest float32 to the exact value and upstream is one
+        # ULP low. Assert against the float64 reference instead, which both
+        # implementations should round to.
+        assert header.dmean == np.float32(vol.mean(dtype=np.float64))
+        assert header.rms == np.float32(vol.std(dtype=np.float64))
 
     def test_stats_are_updated_on_request(self):
         x, y = 4, 3
